refactor: LifeCycle hooks replaced to components from useTimer, fetchSettings renamed to getSettings, <ol> todoCatalog refactored to TodoList component, refactored todoStore methods
