+++ src/store/todoStore.ts
@@ -6,18 +6,17 @@ import { computed, ref } from 'vue';
 export const useTodoStore = defineStore('todo', () => {
   const todos = ref<Todo[]>([]);
 
-  const getBacklogTodos = computed(() => todos.value.filter(todo => todo.status === TODO_STATUS.BACKLOG));
-  const getInProgressTodos = computed(() => todos.value.filter(todo => todo.status === TODO_STATUS.IN_PROGRESS));
-  const getStoppedTodos = computed(() => todos.value.filter(todo => todo.status === TODO_STATUS.STOPPED));
-  const getDoneTodos = computed(() => todos.value.filter(todo => todo.status === TODO_STATUS.DONE));
+  const getTodosByStatus
+   = computed(() => (todoStatus: TODO_STATUS) => todos.value.filter(todo => todo.status === todoStatus));
 
   const getInProgressTodo = computed(() => todos.value.find(todo => todo.status === TODO_STATUS.IN_PROGRESS));
 
   const getTodoIndexById = (todoId: string) => todos.value.findIndex(todo => todo.id === todoId);
 
-  const createTodo = (newTodo: Todo) => {
+  const createTodo = (newTodo: Todo) =>
     todos.value.push({ ...newTodo, id: crypto.randomUUID(), status: TODO_STATUS.BACKLOG });
-  };
+
+  const changeTodoStatus = (todo: Todo, newTodoStatus: TODO_STATUS) => todo.status = newTodoStatus;
 
   const changeTodoRemainingTime = (timeDone: number) => {
     const inProgressTodo = getInProgressTodo;
@@ -34,37 +33,20 @@ export const useTodoStore = defineStore('todo', () => {
     }
   };
 
-  const startTodo = (todo: Todo) => {
-    todo.status = TODO_STATUS.IN_PROGRESS;
-  };
-
   const removeTodo = (todo: Todo) => {
     const todoIndex = getTodoIndexById(todo.id);
     todos.value.splice(todoIndex, 1);
   };
 
-  const doneTodo = (todo: Todo) => {
-    todo.status = TODO_STATUS.DONE;
-  };
-
-  const stopTodo = (todo: Todo) => {
-    todo.status = TODO_STATUS.STOPPED;
-  };
-
   return {
     todos,
-    getBacklogTodos,
-    getInProgressTodos,
-    getStoppedTodos,
-    getDoneTodos,
     getInProgressTodo,
+    getTodosByStatus,
 
     createTodo,
     changeTodoRemainingTime,
     changeTodoOnDrop,
+    changeTodoStatus,
     removeTodo,
-    doneTodo,
-    startTodo,
-    stopTodo,
   };
 });
